New review for CPSC 427 by Mana  (#733)
> Course was composed of a project (60%) and 4 assignments (40%) in
which you need to pass both the project and assignment portion of the
course. Personally, the assignments took too much time away from the
project and the assignment instructions were vague. However, the project
was a blast if you have the right team.
>
> Difficulty: 3.5/5
> Quality: 4/5
> <cite><a href=''>Mana </a>, May 29 2025, course taken during
2025W2</cite>
<details><summary>View YAML for new review</summary>
<pre>
  - author: Mana 
    authorLink: 
    date: 2025-05-29
    review: |
Course was composed of a project (60%) and 4 assignments (40%) in which
you need to pass both the project and assignment portion of the course.
Personally, the assignments took too much time away from the project and
the assignment instructions were vague. However, the project was a blast
if you have the right team.
    difficulty: 3.5
    quality: 4
    sessionTaken: 2025W2

<pre>
</details>This is an auto-generated PR made using:
https://github.com/ubccsss/course-review-worker