New review for CPSC 313 by Mana  (#734)
> Roughly bi-weekly quizzes that allow retakes for an 80% grade
replacement. The final is extremely difficult. Make sure to do well on
the quizzes and it retakes so you can go into the final exam needing a
lower score to pass.
>
> Difficulty: 4.5/5
> Quality: 4.5/5
> <cite><a href=''>Mana </a>, May 29 2025, course taken during
2025W2</cite>
<details><summary>View YAML for new review</summary>
<pre>
  - author: Mana 
    authorLink: 
    date: 2025-05-29
    review: |
Roughly bi-weekly quizzes that allow retakes for an 80% grade
replacement. The final is extremely difficult. Make sure to do well on
the quizzes and it retakes so you can go into the final exam needing a
lower score to pass.
    difficulty: 4.5
    quality: 4.5
    sessionTaken: 2025W2

<pre>
</details>This is an auto-generated PR made using:
https://github.com/ubccsss/course-review-worker