New review for CPSC 404 by Jose Cuervo (#742)
> Not a difficult course at all, got an easy 85% but not a great course
either. Not relevant to the current industrial landscape either. Very
disorganized under Laks. Tuition is expensive, spend that money on a
course that will be worth it. This course is not it.
>
> Difficulty: 2/5
> Quality: 1/5
> <cite><a href=''>Jose Cuervo</a>, Jun 17 2025, course taken during
2024W2</cite>
<details><summary>View YAML for new review</summary>
<pre>
  - author: Jose Cuervo
    authorLink: 
    date: 2025-06-17
    review: |
Not a difficult course at all, got an easy 85% but not a great course
either. Not relevant to the current industrial landscape either. Very
disorganized under Laks. Tuition is expensive, spend that money on a
course that will be worth it. This course is not it.
    difficulty: 2
    quality: 1
    sessionTaken: 2024W2

<pre>
</details>This is an auto-generated PR made using:
https://github.com/ubccsss/course-review-worker