New review for CPSC 427 by nobody (#746)
> If you want to become a game designer, I highly recommand this course.
The team project will give you a solid support on your resume and in
your interview. But if you want to learn modern engines like Unity or
UE5, this might not be a good course. If you already joined this course,
remember to find a good team!!! And I can tell this class is getting
better every year, so if you see my comment from years later, do not be
tricked.
>
> Difficulty: 3.5/5
> Quality: 4.5/5
> <cite><a href=''>nobody</a>, Jun 27 2025, course taken during
2024W2</cite>
<details><summary>View YAML for new review</summary>
<pre>
  - author: nobody
    authorLink: 
    date: 2025-06-27
    review: |
If you want to become a game designer, I highly recommand this course.
The team project will give you a solid support on your resume and in
your interview. But if you want to learn modern engines like Unity or
UE5, this might not be a good course. If you already joined this course,
remember to find a good team!!! And I can tell this class is getting
better every year, so if you see my comment from years later, do not be
tricked.
    difficulty: 3.5
    quality: 4.5
    sessionTaken: 2024W2

<pre>
</details>This is an auto-generated PR made using:
https://github.com/ubccsss/course-review-worker